requirements.txt
pymongo==4.11.3
langchain==0.3.25
langchain-aws==0.2.22
langchain-google-genai==2.1.4
langchain-openai==0.3.16
langgraph==0.4.2
langgraph-checkpoint-mongodb==0.1.3
tiktoken==0.9.0
voyageai==0.3.4
tqdm==4.67.1
Pillow==11.1.0
fastapi
uvicorn
sse-starlette
vercel
python-dotenv

